Comments added

+++ baadalinstallation/baadal/models/admin_model.py
@@ -1,9 +1,13 @@
 # -*- coding: utf-8 -*-
+"""
+admin_model.py: This model has functions to handle database transaction for 
+functionalities that is accessible to users with role of system administrator.
+"""
 ###################################################################################
 # Added to enable code completion in IDE's.
 if 0:
     from gluon import *  # @UnusedWildImport
-    from gluon import db, request
+    from gluon import db, request, session
     from applications.baadal.models import *  # @UnusedWildImport
 ###################################################################################
 
@@ -624,7 +628,7 @@ def get_migrate_vm_details(vm_id):
     return vm_details
 
 
-def get_vm_details(vm_id):
+def get_vm_host_details(vm_id):
     vm_data = db.vm_data[vm_id]
     vm_details = {}
     vm_details['vm_id'] = vm_id
@@ -982,21 +986,21 @@ def get_mail_user_form():
 
 
 
-def add_data_into_affinity(params,vm_details):
-    current_host = vm_details['current_host']
-    current_host = current_host.split("(")
-    current_host = current_host[0]
+def add_data_into_affinity(params, vm_id):
+#     current_host = vm_details['current_host']
+#     current_host = current_host.split("(")
+#     current_host = current_host[0]
     if isinstance(params['affinity_host'], str):
-       host_details = params['affinity_host'].split()
+        host_details = params['affinity_host'].split()
     else :
-       host_details = params['affinity_host'] 
+        host_details = params['affinity_host'] 
     if host_details != None:
-       for host_name in host_details:
-          db(db.vm_data.id == vm_details['vm_id']).update(affinity_flag=1)
-          if db.host_affinity(affinity_host=host_name):
-             db(db.host_affinity.affinity_host == host_name).update(affinity_host=host_name)
-          else :
-             db.host_affinity.insert(vm_id=vm_details['vm_id'], affinity_host=host_name)
+        for host_name in host_details:
+            db(db.vm_data.id == vm_id).update(affinity_flag=1)
+            if db.host_affinity(affinity_host=host_name):
+                db(db.host_affinity.affinity_host == host_name).update(affinity_host=host_name)
+            else :
+                db.host_affinity.insert(vm_id=vm_id, affinity_host=host_name)
     return host_details
 
 
@@ -1004,13 +1008,19 @@ def get_host_details(vm_id):
     host_data = db(db.host_affinity.vm_id == vm_id).select().first()
     host_details = {}
     if host_data != None :
-       host_details['available_hosts'] = dict((host.affinity_host, "%s"%(host.affinity_host.host_name))
-                                         for host in db((db.host_affinity.vm_id == host_data.vm_id)).select())
+        host_details['available_hosts'] = dict((host.affinity_host, "%s"%(host.affinity_host.host_name))
+                                          for host in db((db.host_affinity.vm_id == host_data.vm_id)).select())
     return host_details
    
 
 def reset_host_affinity(vm_id,key):  
     host_data = db(db.host_affinity.affinity_host == key).select().first()
-    vm_details = get_vm_details(vm_id) 
-    db(db.host_affinity.affinity_host==key).delete()
-    db(db.vm_data.id == vm_details['vm_id']).update(affinity_flag=0)
+    vm_details = get_vm_host_details(vm_id)
+    affinity_host = host_data['affinity_host']
+    host = db(db.host.id == affinity_host).select().first()
+    if host.host_name in vm_details['current_host']: 
+        session.flash = "we can not delete this host affinity because currently vm is on this host !!"
+    else :
+        db(db.host_affinity.affinity_host==key).delete()
+        db(db.vm_data.id == vm_details['vm_id']).update(affinity_flag=0)
+        session.flash = 'host affinity removed.'
